pyproject.toml
[project]
name = "blog-generation-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "ipykernel>=6.30.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.69",
    "langchain-groq>=0.3.6",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "langgraph-cli[inmem]>=0.3.6",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "watchdog>=6.0.0",
]
